Fix UART0 clock
The issue was that UART0 had a clock index/selector/whatever of 0, and the
generic Marvell clock driver under drivers/clk/mmp/clk-gate.c only links a clock
defined in the driver and a DT index if the said index is non-zero. Since that
wasn't the case for UART0, the UART0 clock in the driver and device tree would
never get linked and in the DT UART0 would get an empty clock. With this commit,
it is now possible to use `console=ttyS0,115200n8` to get a real console rather
than just earlycon.