tackle ist30xxc_update
Lot of random changes, mostly removals so that I don't have to try to
get it to compile. Hopefully it was not crucial (either it was just a
firmware update, which we can live without for now, or it was firmware
loading, which will cause the touchscreen not to work).