"hide" some "unused" ldos from the kernel so that it doesn't disable them
I remember reading about this somewhere (probably the mainlining.dev
blog; at least this is how I understood it): the
bootloader/SoC/something sets up some/all of the hardware before it
passes control to the kernel. If the ldos will be specified in the DT
but not used anywhere, kernel will switch them off to save power,
because from its point of view they are not needed. However at least
some of them are needed and there is a chance they were set up correctly
by the bootloader/...

Leaving them in seem to disable lots of stuff, including the displaym,
so let's hide them from the kernel so that it doesn't turn them off and
let's handle them later when we know what to use them for in the DT.