Revert "input: 88pm88x-onkey: implement long onkey"
This partially reverts commit a93a647dae7ab8ec32e8868a0fd09ec5c10ef9eb.

Enabling the support seems to have caused the phone to turn on upon a
single press of the onkey, instead of a longer hold, which does not seem
desirable.

However holding the key for 15 seconds while the phone is on still
causes it to shut down immediately. Yet, this is a different behaviour
than with long onkey enabled when this timeout was set to 10 seconds and
the device had powered itself back on right after shutting down (hence
probably the RST mentions in the downstream code).

It might be worth investigating, whether this effect is what was
previously (mistakenly) considered to happen after ~20 seconds of
simultaneous hold of onkey and volume down key (this does still work and
also shuts the device down after 15 seconds, so it seems that indeed the
power down button is likely completely irrelevant here).

Also note that taking out the battery seems to disable long onkey, while
just compiling it out and flashing such kernel to the device after it
had previously run a kernel enabling the long onkey does not.