DONOTMERGE: i2c: pxa: disable i2c unit at end of probe
This gets touch working for xcover3lte. It seems to be a general
problem though. The fix needs to be tested with suspend/resume and pio_xfer.
Might affect resume of cyttsp5 as its sleep mode is not used in downstream.
The mechanism is replicated from pio_xfer.

In downstream the i2c unit is disabled at the end of probe and enabled/disabled
on each transmission. Its done because the i2c bus can be "disturbed" if the
bus device powers on (e.g. touch cyttsp5), or "power up of camera sensor ov2659
may introduce noise on sda, which makes i2c controller busy".
I2C bus recovery in down/upstream via gpio doesn't help.

Lets disable the unit after probe so it gets enabled on first transmission.
This consolidate i2c_pxa_*xfer() implementations as well.