mfd: 88pm88x: hardcode interrupt clearing mode
My current understanding is that this is not device specific but rather
platform (Linux,...) specific. And it seems that here we use
clear-on-write everywhere, so we might as well unconditionally set it --
other platforms would need to parse and handle the DT setting in their
drivers anyway so they can also just go ahead and hardcode it according
to their needs.