input/navpoint: remove driver
This driver does not use the SPI core as it should, instead tampering
with the SSP registers manually. Besides this, it relies on the global
GPIO namespace. Refactoring the driver is almost certainly not worth it
as it was seemingly only implemented in the HP iPAQ hx4700 removed more
than a year ago in d6df7df7ae5a ("ARM: pxa: remove unused board files"),
so let's remove it.

Signed-off-by: Duje Mihanović <duje.mihanovic@skole.hr>