DONOTMERGE: arm64: dts: samsung,coreprimevelte: add rfkill node
Inspired by the downstream DTS. Disabling the thus created rfkill device
does indeed shutdown mmc2, however the kernel does not associate this
with the wireless driver and registers it as its own rfkill device on
which it performs "soft" block. Also, unblocking the rfkill device does
not bring mmc2 back. More work on the integration with the wireless
drivers (mwifiex and bt_mrvl) is needed.