DONOTMERGE: arm64: dts: samsung,coreprimevelte: Raise initrd end to 0x6000000
A recent rework to postmarketos-initramfs made it explode in size,
therefore the current 2624KiB of space assumed for the initramfs just
won't cut it anymore. Fortunately, the bootloader does in fact copy more
than the 2624KiB specified in the cmdline, so let's just increase the
end in the device tree.