clk: mmp: drop mmp_vco_params
The motivation is to cleanly split the dynamic parameters (e. g.
apbs_base) and the static ones (e. g. swcr_offset) so that client
drivers (pxa1908-mpmu in this case) can just list the latter and pass
them to the registration function without having to set up the pll
internals themselves. This is similar to other mmp_register_* functions.