DONOTMERGE: samsung,coreprimevelte: bring up speaker and headphones using downstream code
Bring in the relevant vendor drivers to enable playing sound through the
phone speaker and to headphones. This is a provisional least-effort PoC
solution which will now be step-by-step rewritten and hopefully
upstreamed. It is not suitable for normal use as it contains debug code
and does not offer practical functionality: for instance, some of the
modules do not load automatically (likely due to missing device table),
integration with PulseAudio/PipeWire has not been attemped yet, there is
not headphone jack detection support and alsaucm has to be run every
time before the PCM is opened:

	alsaucm open map\ asoc set _verb HiFi set _enadev Speaker

(or ..._enadev Headphones). For this, the corresponding UCM profile
needs to be placed under .../ucm2/map\ asoc. The profile can be obtained
from the pmaports package for xcover3lte.

Playing sound through the headset has not been attempted as well as it
has not been attempted to use the microphone.

Also beware that most of this really needs to be compiled as modules
because clk-audio-map breaks the boot when compiled built-in because of
the call to clk_prepare_enable which in turn calls readl which fails for
unknown reasons (the prepare_enable call most likely isn't even legal
here and will likely be removed in the rewrite) and the other modules
depend directly or indirectly on the clocks being ready.

Note that the mainline version of the DMA (mmp_tdma) driver would
probably work too, the vendor version was brought in as an attempt to
fix the last problem which seemed to be DMA related (this was solved in
the end by bringing in the downstream version of mmp-sspa/pcm).

Note that the volume can be controlled using the PM860_VOL_SEL_*
controls (highest values until tearing occurs seem to be 70 for the
speaker and up to 80 for the headphones).

Use at your own risk.